CoreTest -> CoreStandardTest
